refactor: migrate card form page
