+++ src/features/card/index.ts
@@ -1,7 +1,8 @@
 export { BackText, type BackTextProps } from "./components/BackText";
 export { Card, type CardProps } from "./components/Card";
+export { CardForm, type CardFormFields, type CardFormProps } from "./components/CardForm";
 export { CardOverlay } from "./components/CardOverlay";
 export { FrontText } from "./components/FrontText";
-export { CardFormContainer } from "./containers/CardFormContainer";
 export { CardViewContainer } from "./containers/CardViewContainer";
+export { useCardFormState } from "./hooks/useCardFormState";
 export { useCardMutations } from "./hooks/useCardMutations";
